requirements.txt
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
telebot==0.0.5

